services/data_scrape/requirements.txt
fastapi==0.70.1
praw==7.5.0
python-decouple==3.5
uvicorn